refactor(cluster): extract shared worker protocol
