+++ packages/cluster/src/index.ts
@@ -21,3 +21,11 @@ export async function startCluster(options: ClusterOptions): Promise<void> {
 export { Master, type MasterOptions, type ClusterOptions, type ClusterHTTPSSecureOptions, type ClusterStartMode };
 
 export * from './error/index.ts';
+export {
+  startAppWorker,
+  startAgentWorker,
+  createProcessWorkerIO,
+  type AppWorkerIO,
+  type AgentWorkerIO,
+  type AppWorkerProtocolOptions,
+} from './worker_protocol/index.ts';
