+++ packages/cluster/src/utils/mode/impl/process/agent.ts
@@ -1,7 +1,6 @@
 import { fork, type ChildProcess, type ForkOptions } from 'node:child_process';
 import { debuglog } from 'node:util';
 
-import { graceful as gracefulExit, type Options as gracefulExitOptions } from 'graceful-process';
 import { sendmessage } from 'sendmessage';
 
 import { ClusterAgentWorkerError } from '../../../../error/ClusterAgentWorkerError.ts';
@@ -18,20 +17,6 @@ export class AgentProcessWorker extends BaseAgentWorker<ChildProcess> {
 
   send(message: MessageBody): void {
     sendmessage(this.instance, message);
-  }
-
-  static send(message: MessageBody): void {
-    message.senderWorkerId = String(process.pid);
-    process.send!(message);
-  }
-
-  static kill(): void {
-    process.exitCode = 1;
-    process.kill(process.pid);
-  }
-
-  static gracefulExit(options: gracefulExitOptions): void {
-    gracefulExit(options);
   }
 }
