+++ packages/cluster/src/worker_protocol/index.ts
@@ -1,0 +1,3 @@
+export { startAppWorker, type AppWorkerIO, type AppWorkerProtocolOptions } from './app.ts';
+export { startAgentWorker, type AgentWorkerIO } from './agent.ts';
+export { createProcessWorkerIO } from './process.ts';
