+++ packages/cluster/tsdown.config.ts
@@ -5,5 +5,6 @@ export default defineConfig({
     index: 'src/index.ts',
     agent_worker: 'src/agent_worker.ts',
     app_worker: 'src/app_worker.ts',
+    worker_protocol: 'src/worker_protocol/index.ts',
   },
 });
